(Chore) Split generator
