--- generators/app/templates/app/static/nlmaps.css
@@ -1,93 +1,0 @@
-.nlmaps-hidden {
-  display:none;
-}
-
-/*** Geolocator control ***/
-.nlmaps-geolocator-control {
-  background-color: #fff;
-  cursor: pointer;
-  box-shadow: 0 1px 5px rgba(0, 0, 0, 0.65);
-  height: 26px;
-  width: 26px;
-  border-radius: 26px 26px;
-  right: .5em;
-  top: .5em;
-}
-
-.nlmaps-geolocator-control img {
-  display: block;
-  box-sizing: border-box;
-  width: 26px;
-  height: 26px;
-  content: none;
-}
-/*myControlEl.className = 'ol-control';*/
-
-
-/*** Geocoder control ***/
-.nlmaps-geocoder-control-container {
-  width: 326px;
-  z-index: 1000000;
-  position: absolute;
-  top: 15px;
-  left: 12px;
-}
-.nlmaps-geocoder-control-search {
-  display: flex;
-  flex-flow: row;
-}
-
-.nlmaps-geocoder-control-container input[type="text"]{
-  flex-grow: 1;
-  padding: 4px 10px;
-
-  border: none;
-  background-color: rgb(255, 255, 255);
-  box-shadow: rgba(0, 0, 0, 0.65) 0px 1px 5px;
-  height: 26px;
-  border-radius: 5px 0px 0px 5px;
-  outline: none;
-}
-.nlmaps-geocoder-control-button {
-  width: 34px;
-  border: none;
-  background-color: rgb(255, 255, 255);
-  box-shadow: rgba(0, 0, 0, 0.65) 0px 1px 5px;
-  height: 34px;
-  border-radius: 0px 5px 5px 0px;
-  outline: none;
-  background-size: cover;
-  cursor: pointer;
-}
-.nlmaps-geocoder-control-button:hover {
-  background-color: #6c62a6;
-  color: #fff;
-}
-.nlmaps-geocoder-control-button:hover path {
-  color: #fff;
-}
-
-
-.nlmaps-geocoder-result-list {
-  padding: 10px 10px 2px 10px;
-  width:300px;
-  background-color: #fff;
-  border-radius: 5px 5px;
-  box-shadow: 0 1px 5px rgba(0, 0, 0, 0.65);
-}
-
-.nlmaps-geocoder-result-list li {
-  cursor: pointer;
-  padding: 5px;
-  list-style-type: none;
-  margin-bottom: 5px;
-}
-
-.nlmaps-geocoder-result-list li:hover {
-  background-color: #6c62a6;
-  color: #fff;
-}
-
-.leaflet-grab.nlmaps-marker-cursor {
-  cursor: pointer
-}
